+++ src/main/kotlin/Config.kt
@@ -7,4 +7,6 @@ object Config : AutoSavePluginConfig("setting") {
     var APPID: String by value()
     var USERID: String by value("miraiqqbot")
     var AUTO_ACCEPT: Boolean by value(false)
+    var WHITELISTS: LongArray by value(LongArray(0))
+    var BLACKLISTS: LongArray by value(LongArray(0))
 }