+++ src/main/kotlin/Utils.kt
@@ -4,10 +4,6 @@ import com.alibaba.fastjson.JSONObject
 import net.mamoe.mirai.contact.Group
 import net.mamoe.mirai.message.data.MusicKind
 import net.mamoe.mirai.message.data.MusicShare
-import net.mamoe.mirai.utils.ExternalResource.Companion.toExternalResource
-import net.mamoe.mirai.utils.ExternalResource.Companion.uploadAsVoice
-import java.io.File
-import java.io.IOException
 import java.io.InputStream
 import java.net.URL
 import java.net.URLEncoder
@@ -42,38 +38,6 @@ class Utils {
             val jsonStr = URL(url).readText()
             val jsonObj = JSONObject.parseObject(jsonStr)
             return jsonObj.getJSONObject("data").getJSONObject("info").getString("text")
-        }
-
-        @Suppress("BlockingMethodInNonBlockingContext")
-        suspend fun tts(text: String, group: Group) {
-            val ffmpeg = File("ffmpeg.exe")
-            val silkEncoder = File("silk_v3_encoder.exe")
-            val ttsURL = "https://fanyi.baidu.com/gettts?lan=zh&spd=5&text=${URLEncoder.encode(text, Charsets.UTF_8)}"
-            val stream = URL(ttsURL).openStream()
-
-            if (ffmpeg.exists() && silkEncoder.exists()) {
-                val mp3 = File("mirai_${System.currentTimeMillis()}.mp3")
-                val pcm = File("mirai_${System.currentTimeMillis()}.pcm")
-                val silk = File("mirai_${System.currentTimeMillis()}.silk")
-
-                mp3.writeBytes(stream.readAllBytes())
-
-                try {
-                    exeCmd("${ffmpeg.absolutePath} -i ${mp3.absolutePath} -f s16le -ar 24000 -ac 1 -acodec pcm_s16le -y ${pcm.absolutePath}")
-                    exeCmd("${silkEncoder.absolutePath} ${pcm.absolutePath} ${silk.absolutePath} -Fs_API 24000 -tencent -quiet")
-                } catch (e: IOException) {
-                    group.sendMessage(e.toString())
-                }
-                val er = silk.toExternalResource()
-                group.sendMessage(er.uploadAsVoice(group))
-                mp3.delete()
-                pcm.delete()
-                stream.close()
-                er.close()
-                silk.delete()
-            } else {
-                group.sendMessage("ffmpeg或silk编码器丢失")
-            }
         }
 
         @Suppress("BlockingMethodInNonBlockingContext")
@@ -127,8 +91,7 @@ class Utils {
                         summary = desc,
                         jumpUrl = jumpURL,
                         pictureUrl = picURL,
-                        musicUrl = musicURL,
-                        brief = "[一位沙雕群友点的歌]$musicName"
+                        musicUrl = musicURL
                     )
                 )
             }