+++ CHANGELOG.md
@@ -5,21 +5,34 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2021-6-26
+### Added
+`WHITELISTS` `BLACKLISTS` 配置项  
+`baike` 命令
+### Changed
+不再需要`chat-command`作为前置
+### Removed
+`test` `gnmsmc` `atall` `tts` `say` 命令
+### Fixed
+`getcoin` 指令出现bug的情况
+
 ## [0.4.0] - 2021-6-19
 ### Added
-getavatar命令
+`getavatar` 命令
 
 ## [0.3.0] - 2021-6-14
 ### Added
-qd cointop music命令
+`qd` `cointop` `music` 命令
 
 ## [0.2.0] - 2021-5-9
 ### Added
-AUTO_ACCEPT配置项 gnmsmc命令
+`AUTO_ACCEPT` 配置项  
+`gnmsmc` 命令
 
 ## [0.1.0] - 2021-5-2
 First release!
 
+[0.5.0]: https://github.com/493505110/MiraiQQBOT/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/493505110/MiraiQQBOT/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/493505110/MiraiQQBOT/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/493505110/MiraiQQBOT/compare/0.1.0...0.2.0